Extract reserve repository mixin
