+++ changelog.d/+reserve-repository-slice.changed.md
@@ -1,0 +1,2 @@
+Moved character reserve persistence into a dedicated repository slice while
+preserving the forum repository facade.
