+++ pyproject.toml
@@ -97,6 +97,7 @@ ignore = [
 "src/elbysodic/db/repositories/materials.py" = ["S608"]
 "src/elbysodic/db/repositories/plot_hooks.py" = ["S608"]
 "src/elbysodic/db/repositories/plotting.py" = ["S608"]
+"src/elbysodic/db/repositories/reserves.py" = ["S608"]
 "src/elbysodic/db/repositories/wanted.py" = ["S608"]
 
 # =============================================================================
